+++ libs/base/src/code/result/own/mod.rs
@@ -1,6 +1,6 @@
-// devela::code::result::own
+// devela_base::code::result::own
 //
-//!
+//! Defines [`Own`].
 //
 
 mod general;
@@ -117,16 +117,16 @@ pub struct Own<S, V> {
 }
 
 mod core_impls {
-    use crate::{ConstDefault, Debug, Display, FmtResult, Formatter, Ordering, Own};
+    use crate::{Debug, Display, FmtResult, Formatter, Ordering, Own};
 
     impl<S: Default, V: Default> Default for Own<S, V> {
         fn default() -> Self {
             Own::new(S::default(), V::default())
         }
     }
-    impl<S: ConstDefault, V: ConstDefault> ConstDefault for Own<S, V> {
-        const DEFAULT: Self = Own::new(S::DEFAULT, V::DEFAULT);
-    }
+    // impl<S: ConstDefault, V: ConstDefault> ConstDefault for Own<S, V> {
+    //     const DEFAULT: Self = Own::new(S::DEFAULT, V::DEFAULT);
+    // }
 
     impl<S: Debug, V: Debug> Debug for Own<S, V> {
         fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {