+++ libs/base/src/code/result/own/value.rs
@@ -1,4 +1,4 @@
-// devela::code::result::own::value
+// devela_base::code::result::own::value
 //
 //!
 //