+++ src/num/int/wrapper/namespace.rs
@@ -81,28 +81,29 @@ mod core_impls {
     impl<T: PartialEq> PartialEq<T> for Int<T> {
         fn eq(&self, other: &T) -> bool { self.0.eq(other) }
     }
+    // TODO: re-enable once moved to [base]
     // with ValueQuant:
-    impl<T: PartialEq> PartialEq<ValueQuant<T, T>> for ValueQuant<Int<T>, Int<T>> {
-        fn eq(&self, other: &ValueQuant<T, T>) -> bool {
-            self.v.eq(&other.v) && self.q.eq(&other.q)
-        }
-    }
-    impl<T: PartialEq> PartialEq<ValueQuant<Int<T>, Int<T>>> for ValueQuant<T, T> {
-        fn eq(&self, other: &ValueQuant<Int<T>, Int<T>>) -> bool {
-            self.v.eq(&other.v.0) && self.q.eq(&other.q.0)
-        }
-    }
-    // with ValueQuant and tuple:
-    impl<T: PartialEq> PartialEq<(T, T)> for ValueQuant<Int<T>, Int<T>> {
-        fn eq(&self, other: &(T, T)) -> bool {
-            self.v.eq(&other.0) && self.q.eq(&other.1)
-        }
-    }
-    impl<T: PartialEq> PartialEq<(Int<T>, Int<T>)> for ValueQuant<T, T> {
-        fn eq(&self, other: &(Int<T>, Int<T>)) -> bool {
-            self.v.eq(&other.0.0) && self.q.eq(&other.1.0)
-        }
-    }
+    // impl<T: PartialEq> PartialEq<ValueQuant<T, T>> for ValueQuant<Int<T>, Int<T>> {
+    //     fn eq(&self, other: &ValueQuant<T, T>) -> bool {
+    //         self.v.eq(&other.v) && self.q.eq(&other.q)
+    //     }
+    // }
+    // impl<T: PartialEq> PartialEq<ValueQuant<Int<T>, Int<T>>> for ValueQuant<T, T> {
+    //     fn eq(&self, other: &ValueQuant<Int<T>, Int<T>>) -> bool {
+    //         self.v.eq(&other.v.0) && self.q.eq(&other.q.0)
+    //     }
+    // }
+    // // with ValueQuant and tuple:
+    // impl<T: PartialEq> PartialEq<(T, T)> for ValueQuant<Int<T>, Int<T>> {
+    //     fn eq(&self, other: &(T, T)) -> bool {
+    //         self.v.eq(&other.0) && self.q.eq(&other.1)
+    //     }
+    // }
+    // impl<T: PartialEq> PartialEq<(Int<T>, Int<T>)> for ValueQuant<T, T> {
+    //     fn eq(&self, other: &(Int<T>, Int<T>)) -> bool {
+    //         self.v.eq(&other.0.0) && self.q.eq(&other.1.0)
+    //     }
+    // }
 
     /* ord*/